refact: dashboard home files and timeline
